Fix for crash associated with TCPClient deletion
Should fix couple of crashes when TCPClient is either trying to reconnect or send message scheduled in an event loop but the HelixUser has already disconnected from the proxy, thus deleting HelixUser object and associated TCPClient. Now TCPClient deletion will wait for 10 seconds if no connection is being made currently in case it is scheduled in the event loop.